requirements.txt
flask==3.1.1
mysql-connector-python==9.4.0
pydantic==2.11.7
werkzeug